modularise rope
